Many changes to colvarbias_reweightaMD.
1. The whole class is splitted into seperated .cpp and .h files;
2. Assign an invalid value to `previous_bin` and avoid use the boolean
   flag `firsttime`;
3. Avoid hiding member variable `output_prefix` from colvarbias;
4. Avoid assigning the loop variable in for loop;
5. Add missing `virtual` in covlarproxy_namd.h;
6. Add doxygen comments to `colvarbias_reweightaMD`;
7. Use `cvm::logn` instead of `std::log`;
8. Change `WriteGradients` to `writePMFGradients` as suggested;
9. Clarify what is `\beta` in the documentation;
10. Cite CZAR correctly;
11. Cleanup some dead code.
