+++ namd/colvars/Make.depends
@@ -105,6 +105,25 @@ obj/colvarbias_histogram.o: \
 	colvars/src/colvarbias.h \
 	colvars/src/colvargrid.h
 	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvarbias_histogram.o $(COPTC) colvars/src/colvarbias_histogram.cpp
+obj/colvarbias_histogram_reweight_amd.o: \
+	obj/.exists \
+	colvars/src/colvarbias_histogram_reweight_amd.cpp \
+	colvars/src/colvarbias_histogram_reweight_amd.h \
+	colvars/src/colvarbias_histogram.h \
+	colvars/src/colvarbias.h \
+	colvars/src/colvar.h \
+	colvars/src/colvarmodule.h \
+	colvars/src/colvars_version.h \
+	colvars/src/colvarvalue.h \
+	colvars/src/colvartypes.h \
+	colvars/src/colvarparse.h \
+	colvars/src/colvarparams.h \
+	colvars/src/colvardeps.h \
+	colvars/src/colvargrid.h \
+	colvars/src/colvarproxy.h \
+	colvars/src/colvarproxy_tcl.h \
+	colvars/src/colvarproxy_volmaps.h
+	$(CXX) $(COLVARSCXXFLAGS) $(COPTO)obj/colvarbias_histogram_reweight_amd.o $(COPTC) colvars/src/colvarbias_histogram_reweight_amd.cpp
 obj/colvarbias_meta.o: \
 	obj/.exists \
 	colvars/src/colvarbias_meta.cpp \
