+++ namd/colvars/src/Makefile.namd
@@ -5,6 +5,7 @@ COLVARSLIB = \
 	$(DSTDIR)/colvarbias_abf.o \
 	$(DSTDIR)/colvarbias_alb.o \
 	$(DSTDIR)/colvarbias_histogram.o \
+	$(DSTDIR)/colvarbias_histogram_reweight_amd.o \
 	$(DSTDIR)/colvarbias_meta.o \
 	$(DSTDIR)/colvarbias_restraint.o \
 	$(DSTDIR)/colvarcomp.o \
