--- namd/tests/library/029_reweightamd/AutoDiff/test.production_amd.dat
@@ -1,193 +1,0 @@
-# 1
-#          3       0.1       190  0
-
- 3.05  0
- 3.15  0
- 3.25  0
- 3.35  0
- 3.45  0
- 3.55  0
- 3.65  0
- 3.75  0
- 3.85  0
- 3.95  0
- 4.05  0
- 4.15  0
- 4.25  0
- 4.35  0
- 4.45  0
- 4.55  6
- 4.65  14
- 4.75  0
- 4.85  0
- 4.95  0
- 5.05  0
- 5.15  0
- 5.25  0
- 5.35  0
- 5.45  0
- 5.55  0
- 5.65  0
- 5.75  0
- 5.85  0
- 5.95  0
- 6.05  0
- 6.15  0
- 6.25  0
- 6.35  0
- 6.45  0
- 6.55  0
- 6.65  0
- 6.75  0
- 6.85  0
- 6.95  0
- 7.05  0
- 7.15  0
- 7.25  0
- 7.35  0
- 7.45  0
- 7.55  0
- 7.65  0
- 7.75  0
- 7.85  0
- 7.95  0
- 8.05  0
- 8.15  0
- 8.25  0
- 8.35  0
- 8.45  0
- 8.55  0
- 8.65  0
- 8.75  0
- 8.85  0
- 8.95  0
- 9.05  0
- 9.15  0
- 9.25  0
- 9.35  0
- 9.45  0
- 9.55  0
- 9.65  0
- 9.75  0
- 9.85  0
- 9.95  0
- 10.05  0
- 10.15  0
- 10.25  0
- 10.35  0
- 10.45  0
- 10.55  0
- 10.65  0
- 10.75  0
- 10.85  0
- 10.95  0
- 11.05  0
- 11.15  0
- 11.25  0
- 11.35  0
- 11.45  0
- 11.55  0
- 11.65  0
- 11.75  0
- 11.85  0
- 11.95  0
- 12.05  0
- 12.15  0
- 12.25  0
- 12.35  0
- 12.45  0
- 12.55  0
- 12.65  0
- 12.75  0
- 12.85  0
- 12.95  0
- 13.05  0
- 13.15  0
- 13.25  0
- 13.35  0
- 13.45  0
- 13.55  0
- 13.65  0
- 13.75  0
- 13.85  0
- 13.95  0
- 14.05  0
- 14.15  0
- 14.25  0
- 14.35  0
- 14.45  0
- 14.55  0
- 14.65  0
- 14.75  0
- 14.85  0
- 14.95  0
- 15.05  0
- 15.15  0
- 15.25  0
- 15.35  0
- 15.45  0
- 15.55  0
- 15.65  0
- 15.75  0
- 15.85  0
- 15.95  0
- 16.05  0
- 16.15  0
- 16.25  0
- 16.35  0
- 16.45  0
- 16.55  0
- 16.65  0
- 16.75  0
- 16.85  0
- 16.95  0
- 17.05  0
- 17.15  0
- 17.25  0
- 17.35  0
- 17.45  0
- 17.55  0
- 17.65  0
- 17.75  0
- 17.85  0
- 17.95  0
- 18.05  0
- 18.15  0
- 18.25  0
- 18.35  0
- 18.45  0
- 18.55  0
- 18.65  0
- 18.75  0
- 18.85  0
- 18.95  0
- 19.05  0
- 19.15  0
- 19.25  0
- 19.35  0
- 19.45  0
- 19.55  0
- 19.65  0
- 19.75  0
- 19.85  0
- 19.95  0
- 20.05  0
- 20.15  0
- 20.25  0
- 20.35  0
- 20.45  0
- 20.55  0
- 20.65  0
- 20.75  0
- 20.85  0
- 20.95  0
- 21.05  0
- 21.15  0
- 21.25  0
- 21.35  0
- 21.45  0
- 21.55  0
- 21.65  0
- 21.75  0
- 21.85  0
- 21.95  0
