+++ src/colvarmodule.cpp
@@ -18,6 +18,7 @@
 #include "colvarbias_abf.h"
 #include "colvarbias_alb.h"
 #include "colvarbias_histogram.h"
+#include "colvarbias_histogram_reweight_amd.h"
 #include "colvarbias_meta.h"
 #include "colvarbias_restraint.h"
 #include "colvarscript.h"
